Allow group_by_vars to be configurable by the user
For now we will use the project column for cross-team cohort analysis
and the batch column for within-team analysis.
This has been implemented to allow for multiple columns to be used for
this correction, though only one is being used at present.
Also adds project to the seurat object metadata.